Separate todo-list into its own component
