+++ src/app.html
@@ -1,18 +1,10 @@
 <template>
+  <require from="./components/todo-list"></require>  
   <h1>Todo list</h1>
 
   <form submit.trigger="addTodo()">
     <input type="text" value.bind="todoDescription">
     <button type="submit" disabled.bind="!todoDescription">Add Todo</button>
   </form>
-
-  <ul>
-    <li repeat.for="todo of todos">
-      <input type="checkbox" checked.bind="todo.done">
-      <span css="text-decoration: ${todo.done ? 'line-through' : 'none'}">
-        ${todo.description}
-      </span>
-      <button click.trigger="removeTodo(todo)">Remove</button>
-    </li>
-  </ul>
+  <todo-list view-model.ref="todoList"></todo-list>
 </template>
