+++ src/app.ts
@@ -1,27 +1,16 @@
-import { Todo } from './interfaces/todo';
+import { TodoList } from "components/todo-list";
 
 export class App {
-  todos: Todo[] = [];
   todoDescription = '';
+  todoList : TodoList;
 
   addTodo() {
     if (this.todoDescription) {
-      this.addOne(this.todoDescription);
+      this.todoList.addOne(this.todoDescription);
       this.todoDescription = '';
     }
   }
 
-  private addOne(todoDescription: string) {
-    this.todos.push({
-      description: todoDescription,
-      done: false
-    });
-  }
-
-  removeTodo(todo: Todo) {
-    let index = this.todos.indexOf(todo);
-    if (index !== -1) {
-      this.todos.splice(index, 1);
-    }
-  }
 }
+
+
